Move parsing of scenario results data to chartdata from scenario-results component to create-scenarios component
